Add draft of a multi-view system for the APC
Each of the three areas (button matrix, vertical buttons, horizontal
buttons+faders) can now be bound separately to multiple plugins. For
each plugin, a multiplexer controls which plugin is in the foreground.
Assignment works for all areas already, but switching only works for
the main button matrix (press the shift button to cycle through the
plugins).
When a plugin goes to the background, the multiplexer saves its
button LED state before restoring the LED state of the plugin that is
coming to the foreground.