Add support for turning buttons off during standby
This adds an option to connect to systemd-logind via DBus. Using the
suspend/wake callback, LaunchPy will blank/unblank the APC mini such
that all lights are turned off while the system is sleeping.

The implementation requires dbus_next to be installed and requires
systemd. If you don't want that, just set the ENABLE_STANDBY_SUPPORT
option to False in settings.py.

The current implementation uses asyncio to implement the DBus monitor;
my understanding of asyncio is limited and this implementation might not
be optimal, especially since it wraps the entire startup/shutdown.
process in asyncio. Contributions welcome.